Don't raise error if there are files to be deleted

+++ pyupdate/update.py
@@ -327,7 +327,7 @@ class UpdateManager:
                     self.download_files(file_dir, required=True)
                 update_details['update'] = [file_path for file_path in db_summary.unique_files_cloud_db] + [file_path for file_path, _, _ in db_summary.bad_files]
             
-            if cloud_config['required_only'] and not update_details['update']:
+            if cloud_config['required_only'] and not update_details['update'] and not update_details['delete']:
                 shutil.rmtree(file_dir)
                 raise NoUpdateError('No files to update. | If you wish to update anyway, set required_only to False in the cloud config.')
